Refactor the top-list
Made a new 'RankedList' component with type as props that can take and render all the lists instead of separate pages for top-movies, top-series, popular-movies, and popular-series. Used the Placeholder component to lazyload the images and provide a placeholder meanwhile. Also updated the classNames and stylings according to the latest changes